refractor - enums added to enum names, moved to separate package
