+++ backend/src/main/java/cz/muni/fi/pa165/pneuservis/backend/entity/tire/TryOut.java
@@ -5,6 +5,9 @@
  */
 package cz.muni.fi.pa165.pneuservis.backend.entity.tire;
 
+import cz.muni.fi.pa165.pneuservis.backend.enums.SeasonEnum;
+import cz.muni.fi.pa165.pneuservis.backend.enums.TireVehicleTypeEnum;
+import cz.muni.fi.pa165.pneuservis.backend.enums.SpeedClassEnum;
 import cz.muni.fi.pa165.pneuservis.backend.dao.TireManagerDaoImpl;
 import java.math.BigDecimal;
 import javax.persistence.EntityManager;
@@ -26,11 +29,11 @@ public class TryOut {
         emf = Persistence.createEntityManagerFactory("default");
         
         Tire t1 = new Tire(new Manufacturer("Continental"), new TireProperties(
-                TireVehicleType.P, 165, 95, 16, 95,  SpeedClass.H, Season.WINTER), 10, new BigDecimal(10.568));
+                TireVehicleTypeEnum.P, 165, 95, 16, 95,  SpeedClassEnum.H, SeasonEnum.WINTER), 10, new BigDecimal(10.568));
         Tire t2 = new Tire(new Manufacturer("Continental"), new TireProperties(
-                TireVehicleType.P, 170, 95, 16, 95,  SpeedClass.H, Season.WINTER), 10, new BigDecimal(11.568));
+                TireVehicleTypeEnum.P, 170, 95, 16, 95,  SpeedClassEnum.H, SeasonEnum.WINTER), 10, new BigDecimal(11.568));
         Tire t3 = new Tire(new Manufacturer("Continental"), new TireProperties(
-                TireVehicleType.P, 165, 95, 16, 95,  SpeedClass.H, Season.WINTER), 8, new BigDecimal(50.568));
+                TireVehicleTypeEnum.P, 165, 95, 16, 95,  SpeedClassEnum.H, SeasonEnum.WINTER), 8, new BigDecimal(50.568));
         
         EntityManager em = emf.createEntityManager();
         TireManagerDaoImpl tm = new TireManagerDaoImpl(em);
