+++ backend/src/main/java/cz/muni/fi/pa165/pneuservis/backend/enums/SeasonEnum.java
@@ -3,18 +3,17 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package cz.muni.fi.pa165.pneuservis.backend.entity.tire;
+package cz.muni.fi.pa165.pneuservis.backend.enums;
 
 /***
  * @author Jakub Palenik, 422453@mail.muni.cz
  */
-public enum TireVehicleType {
-    P("Passenger Car"),
-    LT("Light Truck"),
-    ST("Special Trailer"),
-    T("Temporary");
+public enum SeasonEnum {
+    SUMMER("Summer"),
+    WINTER("Winter"),
+    FOUR_SEASONS("Four seasons");
     
     private final String id;
-    TireVehicleType(String id){this.id = id;}
+    SeasonEnum(String id){this.id = id;}
     public String getValue() {return id;} 
 }
