+++ backend/src/main/java/cz/muni/fi/pa165/pneuservis/backend/enums/TireVehicleTypeEnum.java
@@ -3,17 +3,18 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package cz.muni.fi.pa165.pneuservis.backend.entity.tire;
+package cz.muni.fi.pa165.pneuservis.backend.enums;
 
 /***
  * @author Jakub Palenik, 422453@mail.muni.cz
  */
-public enum Season {
-    SUMMER("Summer"),
-    WINTER("Winter"),
-    FOUR_SEASONS("Four seasons");
+public enum TireVehicleTypeEnum {
+    P("Passenger Car"),
+    LT("Light Truck"),
+    ST("Special Trailer"),
+    T("Temporary");
     
     private final String id;
-    Season(String id){this.id = id;}
+    TireVehicleTypeEnum(String id){this.id = id;}
     public String getValue() {return id;} 
 }
